requirements.txt
cognee
lancedb
kuzu
python-dotenv
pandas
requests
slugify
ffmpeg-python
numpy
yt-dlp
